Refactor to use partial classes.
