+++ MonoGame.Framework/Microsoft.Devices.Sensors.targets
@@ -9,5 +9,7 @@
     <Compile Include="Microsoft\Devices\Sensors\SensorFailedException.cs" />
     <Compile Include="Microsoft\Devices\Sensors\SensorReadingEventArgs.cs" />
     <Compile Include="Microsoft\Devices\Sensors\SensorState.cs" />
+    <Compile Include="Microsoft\Devices\Sensors\Accelerometer.cs" />
+    <Compile Include="Microsoft\Devices\Sensors\Compass.cs" />
   </ItemGroup>
 </Project>
