+++ MonoGame.Framework/MonoGame.Framework.DesktopGL.csproj
@@ -29,6 +29,7 @@
     <Compile Remove="Platform\**\*" />
     <Compile Remove="Properties\**\*" />
     <Compile Remove="Utilities\System.Numerics.Vectors\**\*" />
+    <Compile Remove="Microsoft\**\*" />
   </ItemGroup>
 
   <ItemGroup>
@@ -94,4 +95,5 @@
 
   <Import Project="Platform\OpenGL.targets" />
   <Import Project="Platform\OpenAL.targets" />
+  <Import Project="Microsoft.Devices.Sensors.targets" />
 </Project>
