+++ MonoGame.Framework/MonoGame.Framework.WindowsDX.csproj
@@ -19,6 +19,7 @@
     <Compile Remove="Platform\**\*" />
     <Compile Remove="Properties\**\*" />
     <Compile Remove="Utilities\System.Numerics.Vectors\**\*" />
+    <Compile Remove="Microsoft\**\*" />
   </ItemGroup>
 
   <ItemGroup>
@@ -96,4 +97,5 @@
 
   <Import Project="Platform\DirectX.targets" />
   <Import Project="Platform\XAudio.targets" />
+  <Import Project="Microsoft.Devices.Sensors.targets" />
 </Project>
